pyproject.toml
[project]
name = "local-hybrid-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain",
    "langchain-core",
    "langchain-chroma",
    "langchain-openai",
    "langchain-ollama",
    "langchain-text-splitters",
    "langgraph",
    "graphiti-core",
    "langchain-graphiti",
    "neo4j",
    "rich",
    "python-dotenv",
]
